+++ torchbearer/callbacks/terminate_on_nan.py
@@ -9,6 +9,22 @@ import math
 class TerminateOnNaN(Callback):
     """Callback which montiors the given metric and halts training if its value is nan or inf.
 
+    Example: ::
+
+        >>> import torch.nn
+        >>> from torchbearer import Trial
+        >>> from torchbearer.callbacks import TerminateOnNaN
+
+        # Example Trial which terminates on a NaN, forced by a separate callback. Terminates on the 11th batch since
+        the running loss only updates every 10 iterations.
+        >>> term = TerminateOnNaN(monitor='running_loss')
+        >>> @torchbearer.callbacks.on_criterion
+        ... def force_terminate(state):
+        ...     if state[torchbearer.BATCH] == 5:
+        ...         state[torchbearer.LOSS] = state[torchbearer.LOSS] * torch.Tensor([float('NaN')])
+        >>> trial = Trial(None, callbacks=[term, force_terminate], metrics=['loss'], verbose=2).for_steps(30).run(1)
+        Invalid running_loss, terminating
+
     Args:
         monitor (str): The name of the metric to monitor