+++ CHANGELOG.md
@@ -6,12 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Added ``with_loader`` trial method that allows running of custom batch loaders
+- Added a Mock Model which is set when None is passed as the model to a Trial. Mock Model always returns None. 
+- Added `__call__(state)` to `StateKey` so that they can now be used as losses
 ### Changed
 ### Deprecated
 ### Removed
 - Removed the variational sub-package, this will now be packaged separately
 ### Fixed
 - Fixed a bug where list or dictionary metrics would cause the tensorboard callback to error
+- Fixed a bug where running a trial without training steps would error
 
 ## [0.3.2] - 2019-05-28
 ### Added